requirements.txt
# Core dependencies
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0

# GUI
gradio>=4.0.0

# 3D processing
open3d>=0.17.0
trimesh>=4.0.0
pymeshlab>=2022.2

# Point cloud to mesh conversion
scikit-learn>=1.3.0
scipy>=1.11.0

# SAM dependencies (will be installed from sam-3d-objects repo)
# Note: Follow setup instructions from https://github.com/facebookresearch/sam-3d-objects
# for complete SAM 3D Objects installation

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
opencv-python>=4.8.0
